refactor(api): demote crossfit-wod-backfill to a one-off script
The backfill is intended to be run once per local dev environment (or
ad-hoc when needed), not as part of the recurring cron dispatcher.
Moving it out of src/jobs/ — and out of the JOBS registry — makes that
intent explicit and keeps the production dispatcher minimal.

- apps/api/scripts/backfillCrossfitWods.ts — same per-day loop as
  before, now self-executing (own main() + prisma.$disconnect()).
  Optional CLI arg overrides the day count (defaults to 7).
- apps/api/src/jobs/crossfitWodBackfill.ts — removed.
- apps/api/src/jobs/index.ts — unregistered crossfit-wod-backfill.

The pacificDate.ts helper extracted in the previous commit stays —
it's still legitimately shared between the daily job and the script.

New invocation:
  npx dotenv-cli -e .env -- npx tsx apps/api/scripts/backfillCrossfitWods.ts [days]
